moving imports to /manager/eris-mint
