+++ manager/eris-mint/evm/test/vm_test.go
@@ -8,7 +8,7 @@ import (
 	"testing"
 	"time"
 
-	. "github.com/eris-ltd/eris-db/evm"
+	. "github.com/eris-ltd/eris-db/manager/eris-mint/evm"
 	ptypes "github.com/eris-ltd/eris-db/permission/types"
 	"github.com/eris-ltd/eris-db/txs"
 	. "github.com/tendermint/go-common"
