+++ manager/eris-mint/state/common.go
@@ -2,8 +2,8 @@ package state
 
 import (
 	. "github.com/tendermint/go-common"
-	acm "github.com/eris-ltd/eris-db/account"
-	"github.com/eris-ltd/eris-db/evm"
+	acm "github.com/eris-ltd/eris-db/manager/eris-mint/account"
+	"github.com/eris-ltd/eris-db/manager/eris-mint/evm"
 )
 
 type AccountGetter interface {
