+++ manager/eris-mint/state/permissions_test.go
@@ -8,9 +8,9 @@ import (
 	"testing"
 	"time"
 
-	acm "github.com/eris-ltd/eris-db/account"
+	acm "github.com/eris-ltd/eris-db/manager/eris-mint/account"
 	ptypes "github.com/eris-ltd/eris-db/permission/types"
-	. "github.com/eris-ltd/eris-db/state/types"
+	. "github.com/eris-ltd/eris-db/manager/eris-mint/state/types"
 	"github.com/eris-ltd/eris-db/txs"
 
 	. "github.com/tendermint/go-common"
